+++ electron-react-app-main/app/AppContext.tsx
