+++ electron-react-app-main/app/app.tsx
@@ -1,6 +1,6 @@
-import Home from '@/app/components/home/Home'
+import Home from '@/app/home/Home'
 import './styles/app.css'
-import { AppProvider } from './components/AppContext'
+import { AppProvider } from './AppContext'
 
 export default function App() {
   return (
