+++ electron-react-app-main/app/home/styles.css
