+++ electron-react-app-main/app/types/types.ts
